Program Review and Renew 2
This week you will need to make a minimum of three (3) unit tests per function/method. The three tests should be

1. Basic most common test (what is expected to be normal input)

2. Null edge case

3. Edge case of input (least likely input)

These three tests will have to be built for each class's functions in your code. So for every class, a unit tester class should be built, and for each function, 3 separate unit tests should be built (or a single unit test with 3 test cases input used in it).